C024 Payment is Added.
Penalties implementation is initiated. Required properties are added and a static class 'DayOccurencesFinder' is implemented to identify the dates on which the penalties to be calculated. However, the logic for calculating the MinimumOutstandingForPenalty is yet to implemented and the actual calculation of the Penalties itself has to be implemented. This logic should be implemented as follows:
1. Based on MinimumOutstandingForPenalty =
	a. Sum of last two months maintenance in case maintenance charges changes every month.
	b. 2 x monthly maintenance charges if the maintenance amount is constant.
2. On every date identified in the DayOccurencesFinder ->
	CurrentOutstanding = MaintenanceOnlyOutstanding + Penalty on Last Cutoff Date - Payments Made before the Current Cutoff Date + Other Applicable Penalties.