Main:
Design.ExcelInterop, Domain.Geometry and Domain.Reporting projects are moved to the main solution.

SPro2023ConsoleApp is added to with new working file StaadFile to capture all the functionality required to modify (or even to create in the future) a staad model editor file. This is a work in progress.
Attached properties related to ListBox are being worked out.